+++ src/cli/CliArgs.ts
@@ -1,3 +1,0 @@
-export interface CliArgs {
-    debug?: boolean
-}