+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog][keep-a-changelog], and this project ad
 
 ## [1.1.0] - 2022-07-11
 ### Added
-- Verbose output flag (`-d`, `--debug`).
+- Added verbose output flag (`-v`, `--verbose`).
 ### Changed
 - Improved version output.
 - Added TS type for cli commands.